Add a helper: resetTimestampTo
## Problem
In the hardhat network, `evm_mine` does not allow the timestamp of the
block is lower than the previous block.
This causes errors in some tests because we want to go back to when the
pool starts.

## Solution
Add a test helper `resetTimestampTo`. I divide `advanceTimeto` and
`resetTimeTo` because it is helpful to know when the time may go back.